make responsive player grid a reusable component
